@ fix: resolve MaxIntensity Double/Int mismatch causing silent EEW data loss, improve UI
- Fix critical bug: Wolfx API returns MaxIntensity as Double (e.g. 8.1),
  but models defined it as Int, causing kotlinx.serialization failure.
  All EEW messages were silently discarded by mapNotNull after try-catch.
  Changed MaxIntensity from Int to Double in ScEewResponse, CencEewResponse,
  CqEewResponse, and EewEvent.
- Add Log.w to EewRepository catch block for parse failure visibility
- Add EewSource.description field for source descriptions
- Add SectionHeader composable for grouped settings layout
- Improve ServiceToggleCard: add status icon and active/paused text
- Improve ConnectionStatusChip: add Surface and pulsing indicator dot
- Improve SourceSelector: add description line and radio-button indicator
- Improve ThresholdSettings: add section header, icons, value badges
- Improve SimulationCard: add Science icon and subtitle
- Improve LogExportCard: add BugReport icon and subtitle
- Improve EarthquakeHistoryList: add section header, intensity bar, magnitude badge
- Improve MainScreen: organize settings into Service/Source/Thresholds/Tools groups
- Add 9 new string resources in both en and zh locales
@

+++ app/src/main/java/com/github/mytv/myearthquakealert/data/api/WolfxApiModels.kt
@@ -16,7 +16,7 @@ data class ScEewResponse(
     val Longitude: Double = 0.0,
     @SerialName("Magunitude") val Magunitude: Double = 0.0,
     val Depth: Double? = null,
-    val MaxIntensity: Int = 0,
+    val MaxIntensity: Double = 0.0,
 )
 
 @Serializable
@@ -32,7 +32,7 @@ data class CencEewResponse(
     val Longitude: Double = 0.0,
     val Magnitude: Double = 0.0,
     val Depth: Double? = null,
-    val MaxIntensity: Int = 0,
+    val MaxIntensity: Double = 0.0,
 )
 
 @Serializable
@@ -63,7 +63,7 @@ data class CqEewResponse(
     val Longitude: Double = 0.0,
     val Magnitude: Double = 0.0,
     val Depth: Double? = null,
-    val MaxIntensity: Int = 0,
+    val MaxIntensity: Double = 0.0,
 )
 
 @Serializable
